+++ app/(main-layout)/allProjects/page.tsx
@@ -1,20 +1,26 @@
+import { Metadata } from 'next';
 import ProjectsTable from '@/components/ProjectsTable';
 import CreateProjectButton from '@/components/CreateProjectButton';
-// import { currentUser } from '@clerk/nextjs/server';
-import { checkUser } from '@/lib/checkUser';
-import { isEditorOrAbove } from '@/lib/canEdit';
+import getUserInfo from '@/lib/getUserInfo';
+
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: 'All Projects | Argus',
+    description: 'User &rsquo;s own projects',
+  };
+}
 
 const AllProjectsPage = async () => {
-  // const clerkUser = (await currentUser()) || null;
-  // const userId = clerkUser ? clerkUser.id || null : null;
-  const user = await checkUser();
-  const isEditor = await isEditorOrAbove();
+  const {
+    user,
+    permissions: { isEditorOrAbove, canPrint },
+  } = await getUserInfo();
 
   return (
     <>
       <h1 className="h2">All Projects</h1>
-      {isEditor && <CreateProjectButton />}
-      <ProjectsTable limitToUser={false} user={user} />
+      {isEditorOrAbove && <CreateProjectButton />}
+      <ProjectsTable limitToUser={false} user={user} canPrint={canPrint} />
     </>
   );
 };