+++ app/(main-layout)/createProject/page.tsx
@@ -1,23 +1,27 @@
+import { Metadata } from 'next';
 import ProjectForm from '@/components/ProjectForm';
-import { checkUser } from '@/lib/checkUser';
 import { createProject } from '@/app/actions/projectActions';
-import { isEditorOrAbove } from '@/lib/canEdit';
+import getUserInfo from '@/lib/getUserInfo';
 import { unauthorized } from 'next/navigation';
 
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: 'Create Project | Argus',
+    description: 'Create a new project',
+  };
+}
+
 export default async function CreateProjectPage() {
-  const currentUser = await checkUser();
-  // const currentUser: User | null = await checkUser();
-  const isEditor = await isEditorOrAbove();
+  const {
+    user,
+    permissions: { isEditorOrAbove },
+  } = await getUserInfo();
   const basePath = process.env.NEXT_PUBLIC_APP_BASEPATH ?? '/';
 
-  if (isEditor && currentUser != undefined) {
+  if (isEditorOrAbove && user != undefined) {
     return (
       <>
-        <ProjectForm
-          user={currentUser}
-          action={createProject}
-          basePath={basePath}
-        />
+        <ProjectForm user={user} action={createProject} basePath={basePath} />
       </>
     );
   } else {