+++ app/(main-layout)/editEntry/[id]/page.tsx
@@ -1,13 +1,38 @@
+import { Metadata } from 'next';
 import { EntryWithItems } from '@/types/EntryWithItems';
 import { bibHoldings } from '@/app/actions/almaSearch';
 import getEntryById from '@/app/actions/getEntryById';
 import { CondensedBibHoldings } from '@/types/CondensedBibHoldings';
 import BibResultsWrapper from '@/components/BibResultsWrapper';
-import { nonOwnerEditor } from '@/lib/canEdit';
+import getUserInfo from '@/lib/getUserInfo';
 import NonOwnerAlert from '@/components/NonOwnerAlert';
 import ProjectButtons from '@/components/ProjectButtons';
 import ProjectMetadata from '@/components/ProjectMetadata';
 import { getProject } from '@/app/actions/projectActions';
+
+type MetadataProps = {
+  params: Promise<{ id: string }>;
+};
+
+export async function generateMetadata({
+  params,
+}: MetadataProps): Promise<Metadata> {
+  const { id } = await params;
+  const {
+    data: existingEntry,
+    error: existingEntryError,
+  }: { data?: EntryWithItems; error?: string } = await getEntryById(id);
+  if (existingEntry !== undefined) {
+    return {
+      title: `${existingEntry.itemTitle} | Argus`,
+      description: `Existing Item Page: ${existingEntry.itemTitle}`,
+    };
+  } else {
+    return {
+      title: 'Item Not Found | Argus',
+    };
+  }
+}
 
 export default async function EditEntryPage({
   params,
@@ -26,7 +51,10 @@ export default async function EditEntryPage({
   const { project } = await getProject({ id: projectId.toString() });
 
   const mmsId = existingEntry?.almaId ?? '';
-  const { nonOwnerAlert, canEditBool } = await nonOwnerEditor(projectId);
+  const {
+    permissions: { canEdit, canPrint, nonOwnerEditor },
+  } = await getUserInfo(projectId);
+
   const {
     data: holdingsData,
     error: holdingsError,
@@ -38,12 +66,13 @@ export default async function EditEntryPage({
   }
   return (
     <>
-      {nonOwnerAlert && <NonOwnerAlert />}
+      {nonOwnerEditor && <NonOwnerAlert />}
       <h1 className="h2">
         Editing: <i>{holdingsData && holdingsData.bib_data.title}</i>
       </h1>
       <ProjectButtons
-        canEdit={canEditBool}
+        canEdit={canEdit}
+        canPrint={canPrint}
         onPage="editEntry"
         projectId={projectId}
         divClass="mb-2"
@@ -55,17 +84,8 @@ export default async function EditEntryPage({
         holdingsData={holdingsData}
         actionType="edit"
         existingEntry={existingEntry}
-        isEditor={canEditBool}
+        isEditor={canEdit}
       />
     </>
   );
-
-  //   return (
-  //     <>
-
-  //       <BibEntryComponent entry={} />
-  //       <pre>{JSON.stringify(holdingsData, null, 2)}</pre>
-  //     </>
-  //   );
-  // }
 }