+++ app/(main-layout)/editProject/[id]/page.tsx
@@ -1,17 +1,27 @@
+import { Metadata } from 'next';
 import ProjectForm from '@/components/ProjectForm';
 import { checkUser } from '@/lib/checkUser';
 import { updateProject } from '@/app/actions/projectActions';
 import { getProject } from '@/app/actions/projectActions';
-// import { Project } from '@prisma/client';
-import canEdit, { nonOwnerEditor } from '@/lib/canEdit';
+import getUserInfo from '@/lib/getUserInfo';
 import { redirect } from 'next/navigation';
 import { unauthorized } from 'next/navigation';
 import logger from '@/lib/logger';
 import NonOwnerAlert from '@/components/NonOwnerAlert';
+import ProjectMetadata from '@/components/ProjectMetadata';
+import ProjectButtons from '@/components/ProjectButtons';
 
 interface EditProjectPageProps {
   params: Promise<{ id: string }>;
 }
+
+export async function generateMetadata(): Promise<Metadata> {
+  return {
+    title: 'Edit Project Details | Argus',
+    description: 'Update project name and notes',
+  };
+}
+
 export default async function EditProjectPage({
   params,
 }: EditProjectPageProps) {
@@ -19,10 +29,11 @@ export default async function EditProjectPage({
   const { id } = await params;
   const response = await getProject({ id });
   const project = response.project;
-  const isEditor = await canEdit(id);
-  const { nonOwnerAlert } = await nonOwnerEditor(parseInt(id, 10));
+  const {
+    permissions: { isOwner, isOwnerish },
+  } = await getUserInfo(id);
 
-  if (!isEditor) {
+  if (!isOwnerish) {
     redirect(`/project/${id}`); // go to non-edit version of project page
   }
 
@@ -31,7 +42,16 @@ export default async function EditProjectPage({
   if (currentUser != undefined) {
     return (
       <>
-        {nonOwnerAlert && <NonOwnerAlert />}
+        {/* this NonOwnerAlert has different conditions from others because the only one who should normally edit this page is the owner. */}
+        {!isOwner && <NonOwnerAlert />}
+        <h1 className="h2">Edit Project Details</h1>
+        {project && <ProjectMetadata project={project} />}
+        <ProjectButtons
+          projectId={parseInt(id)}
+          divClass="mb-3"
+          canAssignCoEditors={isOwnerish}
+          onPage="edit-project-details"
+        />
         <ProjectForm
           user={currentUser}
           action={updateProject}