+++ components/UserEditForm.tsx
@@ -1,5 +1,6 @@
 'use client';
-import { useState } from 'react';
+import React from 'react';
+import { useEffect, useState } from 'react';
 import updateUser from '@/app/actions/updateUser';
 import { toast } from 'react-toastify';
 import { User, Role, UserAffiliation, UserStatus } from '@prisma/client';
@@ -11,13 +12,12 @@ import {
   InputGroup,
 } from 'react-bootstrap';
 
-// import { User } from '@/types/User';
-// import { revalidatePath } from 'next/cache';
 interface pageProps {
   user: User;
   actorIsSuperAdmin: boolean;
 }
 export default function UserEditForm({ user, actorIsSuperAdmin }: pageProps) {
+  const [name, setName] = useState(user.name);
   const [role, setRole] = useState(user.role);
   const validRoles = Object.values(Role);
   // type Role = (typeof validRoles)[number];
@@ -25,15 +25,22 @@ export default function UserEditForm({ user, actorIsSuperAdmin }: pageProps) {
   const validStatuses = Object.values(UserStatus);
   const [affiliation, setAffiliation] = useState(user.affiliation);
   const validAffiliations = Object.values(UserAffiliation);
+  const [printSlips, setPrintSlips] = useState(user.printSlips);
   // type Role = (typeof validRoles)[number];
 
+  const handleNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setName(e.target.value);
+  };
+
   const handleChange =
-    (targetField: 'role' | 'status' | 'affiliation') =>
+    (targetField: 'role' | 'status' | 'affiliation' | 'printSlips') =>
     (e: React.ChangeEvent<HTMLSelectElement>) => {
       // console.log(`changing target: ${targetField}, ${e.target.value}`);
       switch (targetField) {
         case 'role':
           setRole(e.target.value as Role);
+          ['admin', 'superadmin'].includes(e.target.value) &&
+            setPrintSlips(true);
           break;
         case 'affiliation':
           setAffiliation(e.target.value as UserAffiliation);
@@ -41,21 +48,28 @@ export default function UserEditForm({ user, actorIsSuperAdmin }: pageProps) {
         case 'status':
           setStatus(e.target.value as UserStatus);
           break;
+        case 'printSlips':
+          setPrintSlips((e.target.value.toLowerCase() === 'true') as boolean);
+          break;
       }
     };
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     const updatedUser = await updateUser(user.id, {
+      name: name,
       role: role as Role,
       status: status as UserStatus,
       affiliation: affiliation as UserAffiliation,
+      printSlips: (printSlips ||
+        role == 'admin' ||
+        role == 'superadmin') as boolean,
     });
     if (updatedUser.error) {
       console.error('Error updating user:', updatedUser.error);
       return;
     }
-    toast.success('User role updated successfully');
+    toast.success('User updated successfully');
   };
 
   const statusPulldown = validStatuses.map((r) => (
@@ -73,15 +87,26 @@ export default function UserEditForm({ user, actorIsSuperAdmin }: pageProps) {
       None
     </option>
   );
-  // statusPulldown.unshift(
-  //   <option key="none" value="">
-  //     None
-  //   </option>
-  // );
+
   return (
     <Form onSubmit={handleSubmit}>
-      <InputGroup>
-        <FormLabel htmlFor="role">Role</FormLabel>
+      <InputGroup className="mb-2 d-flex align-items-center">
+        <FormLabel htmlFor="name" className="me-2">
+          Name
+        </FormLabel>
+        <Form.Control
+          type="text"
+          id="name"
+          defaultValue={name}
+          onChange={(e) =>
+            handleNameChange(e as React.ChangeEvent<HTMLInputElement>)
+          }
+        />
+      </InputGroup>
+      <InputGroup className="mb-2 d-flex align-items-center">
+        <FormLabel htmlFor="role" className="me-2">
+          Role
+        </FormLabel>
         <FormSelect
           id="role"
           value={role ?? ''}
@@ -97,8 +122,10 @@ export default function UserEditForm({ user, actorIsSuperAdmin }: pageProps) {
           )}
         </FormSelect>
       </InputGroup>
-      <InputGroup>
-        <FormLabel htmlFor="affiliation">Affiliation</FormLabel>
+      <InputGroup className="mb-2 d-flex align-items-center">
+        <FormLabel htmlFor="affiliation" className="me-2">
+          Affiliation
+        </FormLabel>
         <FormSelect
           id="affiliation"
           value={affiliation ?? ''}
@@ -108,8 +135,10 @@ export default function UserEditForm({ user, actorIsSuperAdmin }: pageProps) {
             affiliationPulldown}
         </FormSelect>
       </InputGroup>
-      <InputGroup>
-        <FormLabel htmlFor="status">Status</FormLabel>
+      <InputGroup className="mb-2 d-flex align-items-center">
+        <FormLabel htmlFor="status" className="me-2">
+          Status
+        </FormLabel>
         <FormSelect
           id="status"
           value={status ?? ''}
@@ -118,7 +147,20 @@ export default function UserEditForm({ user, actorIsSuperAdmin }: pageProps) {
           {statusPulldown.unshift(blankPullDownOption) && statusPulldown}
         </FormSelect>
       </InputGroup>
-      <Button className="btn btn-primary" type="submit">
+      <InputGroup className="mb-2 d-flex align-items-center">
+        <FormLabel htmlFor="printSlips" className="me-2">
+          Print Slips permissions
+        </FormLabel>
+        <FormSelect
+          id="printSlips"
+          value={printSlips.toString() ?? 'false'}
+          onChange={handleChange('printSlips')}
+        >
+          <option value="false">False</option>
+          <option value="true">True</option>
+        </FormSelect>
+      </InputGroup>
+      <Button className="btn btn-primary mt-4" type="submit">
         Save Changes
       </Button>
     </Form>