+++ types/ArgusUserInfo.ts
@@ -1,0 +1,6 @@
+import { ArgusPermissions } from './ArgusPermissions';
+import { User } from '@prisma/client';
+export interface ArgusUserInfo {
+  permissions: ArgusPermissions;
+  user?: User;
+}